% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{colouriseBlack}
\alias{colouriseBlack}
\alias{colouriseBlue}
\alias{colouriseGreen}
\alias{colouriseRed}
\alias{colouriseYellow}
\alias{colourizeBlack}
\alias{colourizeBlue}
\alias{colourizeGreen}
\alias{colourizeRed}
\alias{colourizeYellow}
\title{colourise}
\usage{
colouriseBlack(message)

colouriseBlue(message)

colouriseGreen(message)

colouriseYellow(message)

colouriseRed(message)

colourizeBlack(message)

colourizeBlue(message)

colourizeGreen(message)

colourizeYellow(message)

colourizeRed(message)
}
\arguments{
\item{message.}{An arbitrary string.}
}
\value{
a string.
}
\description{
add colour to a string.
}

