% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{colourise}
\alias{colourise}
\title{colourise}
\format{\preformatted{List of 5
 $ black :function (message)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 3 47 3 17 17 41 47
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x26166f8> 
 $ blue  :function (message)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 3 47 3 17 17 41 47
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x26166f8> 
 $ green :function (message)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 3 47 3 17 17 41 47
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x26166f8> 
 $ red   :function (message)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 3 47 3 17 17 41 47
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x26166f8> 
 $ yellow:function (message)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 41 3 47 3 17 17 41 47
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x26166f8> 
}}
\usage{
colourise
}
\description{
colourise
}
\keyword{datasets}

